requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Database
sqlalchemy==2.0.43
aiosqlite==0.21.0
alembic==1.14.0

# HTTP Client
aiohttp==3.12.15
httpx==0.28.0

# Environment and Configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.3.0

# Async Support
greenlet==3.2.4
asyncio-mqtt==0.16.1

# Email and Notifications
aiosmtplib==3.0.1
jinja2==3.1.5

# Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Testing
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.12.0

# Production
gunicorn==22.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
